bigfin: Corrected the size field while rbd creation
Bug-URL: https://bugzilla.redhat.com/1361092

+++ provider/block_devices.go
@@ -146,7 +146,7 @@ func createBlockStorage(
 
 	// Create the block device image
 	sizeMBs := utils.SizeFromStr(blockDevice.Size) / 1024
-	cmd := fmt.Sprintf("rbd create %s --cluster %s --size %d --pool %s", blockDevice.Name, clusterName, sizeMBs, backingStorage)
+	cmd := fmt.Sprintf("rbd create %s --cluster %s --size %d --pool %s", blockDevice.Name, clusterName, uint64(sizeMBs), backingStorage)
 	ok, _, err := cephapi_backend.ExecCmd(mon, clusterId, cmd, ctxt)
 	if err != nil || !ok {
 		utils.FailTask(fmt.Sprintf("Creation of block device failed on cluster: %s", clusterName), fmt.Errorf("%s - %v", ctxt, err), t)
@@ -348,7 +348,7 @@ func (s *CephProvider) ResizeBlockDevice(req models.RpcRequest, resp *models.Rpc
 					return
 				}
 				t.UpdateStatus("Resizing block device")
-				cmd := fmt.Sprintf("rbd resize --image %s --cluster %s --size %d --pool %s", blockDevice.Name, cluster.Name, sizeMBs, storage.Name)
+				cmd := fmt.Sprintf("rbd resize --image %s --cluster %s --size %d --pool %s", blockDevice.Name, cluster.Name, uint64(sizeMBs), storage.Name)
 				ok, _, err := cephapi_backend.ExecCmd(monnode.Hostname, *cluster_id, cmd, ctxt)
 				if err != nil || !ok {
 					utils.FailTask(fmt.Sprintf("Resizing of block device failed on cluster: %s", *cluster_id), fmt.Errorf("%s - %v", ctxt, err), t)
